move modelinfo to the arc dialect
