+++ lib/CMakeLists.txt
@@ -13,5 +13,4 @@ add_subdirectory(Reduce)
 add_subdirectory(Scheduling)
 add_subdirectory(Support)
 add_subdirectory(Target)
-add_subdirectory(Tools)
 add_subdirectory(Transforms)
