+++ lib/Dialect/Arc/CMakeLists.txt
@@ -3,6 +3,7 @@ set(CIRCT_Arc_Sources
   ArcFolds.cpp
   ArcOps.cpp
   ArcTypes.cpp
+  ModelInfo.cpp
 )
 
 set(LLVM_OPTIONAL_SOURCES
