--- lib/Tools/Arcilator/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_circt_library(CIRCTArcilatorLib
-  ModelInfo.cpp
-
-  LINK_LIBS PUBLIC
-  CIRCTArc
-  )
