--- lib/Tools/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectory(Arcilator)
